tests/test_project/requirements.txt
pytest==2.3.4
pytest-cov==1.6
pytest-django==2.3.0
pytest-flakes==0.1
pytest-pep8==1.0.4
